ingestion_service/requirements.txt
# Add your project dependencies here
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pydantic
hl7apy
pydicom
python-docx
PyMuPDF  # for PDFs
celery==5.3.6
redis==5.0.3
python-multipart
python-magic


